Multiplexer fixes

+++ gym_multiplexer/real_multiplexer.py
@@ -1,5 +1,6 @@
 import random
 
+import numpy as np
 from gym.spaces import Box, Discrete
 
 from .multiplexer import Multiplexer
@@ -10,7 +11,9 @@ class RealMultiplexer(Multiplexer):
   def __init__(self, control_bits=3, threshold=.5) -> None:
     super().__init__(control_bits)
     self.threshold = threshold
-    self.observation_space = Box(low=0, high=1, shape=(self._observation_string_length, ))
+    self.observation_space = Box(low=0, high=1,
+                                 shape=(self._observation_string_length, ),
+                                 dtype=np.float32)
     self.action_space = Discrete(2)
 
   def _generate_state(self):
